DRY
Abstracted SQL functions into SPM_SQL module and made main more DRY by adding a logging function rather than writing to logs seperately
